feat: Header component, Navbar component, Icon component/ style: Navbar component's styles.
For the header and navbar, I made 2 components, Header and Navbar components. Also for the main icon of the website, I made a separate component to make working with it easier.
For better theme and color management, I installed color-mode module for Nuxt, and for better icon management and accessing icons as components, I installed Nuxt/Icon module. I wanted to use the collection of icons included in the project design itself with the help Icon module, but it needed some server tweaking that I could not do, hence the icon object in nuxt.config.ts file.
Since I decided to use tailwind as the styling framework, I made tailwind.css and tailwind.config.js files (manually); styles are loaded and further configurations for the project were also made. To load the fonts, since Google Fonts didn't include woff2/woff files, I used the CDN and links to load them in the project. Hence, the head links in the nuxt.config.ts file.