feat: router setup, C-Button component / fix: Navbar logical structure and menu / style: Hero component's styling
The router for different pages is now setup and ready to use. So many complications happened and it was mostly because of App.vue file not being named correctly (instead of using A for App, it used a), and case sensitiveness of pages directory. best way to name your pages' component is by kebab-case.
C-Button component which stands for Custom Button, is a new addition; it's a NuxtLink tag, with the additional styles to look like a button. since most of the buttons in the project use the same styles, I decided to make a component out of it.
Links used in the Navbar component was not standard; I replaced the normal a links with more optimized NuxtLink tags. Also added the path property to the array of menuItems to use as the route.
The Style of the Hero component was also improved; The left part of this section is almost complete.