feat: Hero section's status component, installed axios for API requests / fix: container's breakpoints
For the status component in the hero section, I composed an API to get the information; for that, I installed Axios and used to make the request, get the info, and put it on display.
Container's breakpoints broke again; did some debugging and now they seem to be working. However it's not entirely matched with the design or the live preview.
For the part of the hero section, which is created by using a gif file, I originally planned to use NuxtImg component, but since it didn't provide any benefit in this context, I just used a normal img tag. Originally the gif file was in the public folder, but it caused a lot of issues; so I moved to assets folder. Also created a component for it.
Ultimately, the Hero section is complete for now. I may change or tweak it in the future if necessary.