fix: container's max-width on breakpoints / refactor: button component / feat: Main, TrendingCollection, SectionHeader, TrendCollection components
Container still had some issues with what was shown in the live preview; Hard-coded the max-width of the container based on the live preview, and now it matches.
Styles for the CButton component was also overhauled. It's more flexible and has a better reusability.
For the main part of the body, in which all sections are, I made the Main.vue component. For section's headings made the SectionHeader component.
The Trending section is made of 3 structures, turned into a component called TrendCollection, for better reusability.