feat: categories section complete / refactor-style: hover effect for elements
The categories section is complete. The problem with the height of the cards was caused by row template configuration of the grid display; removing it solved the problem while maintaining the desired final results.
The hover effect for different parts of the project, such as TrendCollection component, CreatorCard component and Categories component, was so frequently used that making it into a class component seemed reasonable; now "hover-class" represents the styles necessary for this hover effect and this class will replace the overuse of the same styles from now on.