feat: Discover section, Fetching NFTs API, ArtistProfile component
The Discover section's structure is finished (almost); there is no dynamic data yet, but the api to fetch the data is there and working. NFTinfoCard component is used in two variants, lighter and darker; You could change it using the variant prop (lighter or darker are the only accepted values for this prop).
A new component for profile picture of nft's artist was created.