feat: Discover section; Pinia module added / Refactor: creators' number on different breakpoints
Discover section is done. The data for this section is fetched from Nuxt's built-in API capabilities; NFTs' data is read from the JSON file and sent to the client by the API.

Pinia module is another new addition to the project. For better state-management and also educational purposes.

Creators' number, and Discover section's number of shown items, differ in each breakpoint; therefore I came up with this solution to programmatically figure out the width of the window, and then change the number of displayed items, using a Pinia store to store this "window width" and use it across any components.