feat: Highlight section, artists and NFTs relations with IDs / fix: artist information in Discover section was missing
The Highlight section is complete. The countdown is made using a composable; it takes a date and returns the days, hours, minutes and seconds left to the given time. Also, it provides a "isEnded" flag to display the appropriate content based on it (didn't use it here though).

Using AI assistance, I managed to link the artists with NFTs. Now each NFT has a "creatorID", which is linked to an artist from artists.json file.

The artist information of the items in Discover section is now being fetched, using the new connection between artists and NFTs; The creator's id is being passed to an api, and it returns the information object for that creator. Several other changes were needed to make this process smoother :) But it works ! :)))