feat: How it works & Subscription section implementation / refactor: layout implementation
How it works and Subscription section is implemented.

Since the site's header and footer is shared across all pages, layout implementation is a necessity. Header and Footer components are moved to default.vue file to be used as the default shared components. All the components that used to be in Main.vue file were moved to index.vue file (they belonged to the home page).