feat: Footer, MobileNavMenu implementation / fix: linkTo prop for CButton component
The Footer section is now complete; It's shared among all routes.

Navbar is now separated; for both desktop and mobile. Mobile navbar though is still in progress and not complete

Some section that CButton component was used, had some problems with linkTo prop; it's a required prop but it wasn't passed to the component in sections such as TrendCollection.vue, Header, Discover, etc.