feat: change of default layout / fix: problem with window's width store
The default layout is changed because most of the pages are under development and not ready, I changed the default layout to show a page for all routes; and for the Home page, which is ready, I used the home layout.

The width of the window is used in both the Discover section and Navbar; so that according to that, elements (such as desktop navbar or how many NFT cards are shown) could dynamically appear or disappear. Therefore, I implemented a resizing event into the store for making the process easier.