feat: Login validation error handling, registration and adding new user completion / refactor:
Error handling of the log-in form validation is now functional. Each field is evaluated by ZOD and the schema defined for it. After the validation process, errors are converted into an array, go through a loop, and then shown by toast notifications.

The new user gets added to the artists.json file now; After the operation, the user is faced with the option to also check their account information and complete the rest of the necessary information. Although the user panel is not formed, therefore user is redirected to the home page for now.

A custom input element (CInput.vue) was used in FormInput component for the input; but due to complications caused by it for emitting events and v-model, it was replaced by a normal input element.

p.s: there is a bug regarding the redirection of user to the homepage, after registration is complete and user is prompted with the question to check their panel.