feat: custom table grid for ranking page, displaying user's username instead of sign-up button, new page for user's info / refactor: window store optimation
The ranking page used a unique grid for spacing the content of each row (the header and each row for each creator) which was quite hard to implement; but I cracked it :) Cracked it WIDE OPEN! :)) During which I had to add some custom sized and spacing to the tailwind.config.js file... so that's that. It's 2 in the morning and I'm dying of sleep :)

Now when you log in, your username and email get stored in the userAuth store, and the sign-up button's caption changes to user's username.

Also, window store is updated; Now the state is initialized right when the app is mounted, and can be used across the app. This fixes the issues with Navbar; Navbar is supposed to change form when going to tablet, from desktop, but it was janky and broken, but not anymore :). Some other sections that used the window store also had to be modified to fit the new update.

A new page for displaying user's info and details is created (for now it just displays the username:).