feat: Tab filtering system for Marketplace and Rankings pages / refactor: README file update, API update
The tabs in Rankings and Marketplace pages are now functional, although Rankings page needs further tweaking and developement.

The project's README file is updated, for a better, more aesthetically pleasing experience.

APIs used throughout the project, are updated, from using "readFile" method, to "$fetch"; It used to use a combination of Node's readFile method and local file management system, to read the content form json files, and return them to the client. This caused a lot of problems in production and deployment; therefore, it's switched to using Nuxt's built-in "$fetch" method, to get the data as a HTTP request from the server when it's running, and then display send them to the client. It's less complicated and works better. Also, deployment happened on Liara's Node server.