feat: Artist's detail page, artist status information added to artists.json file / refactor: status component compatibility with artist's status
The information regarding artist's status, such as followers count, nfts sold, and sales volume, is now embedded inside artists.json file and is retrieved during the API call with the rest of the info about the artist.  This procedure caused some incompatibilities with the type initially created for Artist's information object, in Global.ts file; Hence the changes in the mentioned file to resolve the issues.

The StatusBox component was created originally for the status in Hero section of the Home page;  Therefore, some incompatibilities surfaced when used in artist's detail page.  The issues were addressed and dealt with, such as the data types passed into the component, and some small styles.