feat: rankings page now is connected to the artist's information, NFTs' detail page is now under development, clicking on NFT card's title or image, it takes you to NFT details page / fix: the bug with fetching artist's NFTs' information / refactor: a new script for exposing the host to local devices was added to package.json file
The ranking page contains the information of all artists in the website; now when you click on the area where the artist's profile picture and username is, it redirects you to the artist's detail page.
The NFTs' detail page was empty, now it's holding the placeholder component ( UnderDevelop.vue ).

There was a bug with useFindNFTById.ts file that was returning the wrong information; it's fixed.